+++ CHANGELOG.md
@@ -1,3 +1,12 @@
+## v6.1.0
+October 16 2023
+
+#### Added
+- Add crypto module that allows configure SDK to encrypt and decrypt messages.
+
+#### Fixed
+- Improved security of crypto implementation by adding enhanced AES-CBC cryptor.
+
 ## v6.0.1
 May 18 2023